Separates SSA count from symbol id
Before to generate SSA, the symbol id was renamed to
a different symbol with the original symbol appended
with the count. For example, "a" would be renamed to
"a1". This makes it difficult to recover the original
symbol. It also generates a lot of unnecessary temporaries.

Now, names are represented by a tuple of symbol id and count.
Initially, the count is -1, and will be updated after the
SSA renaming pass. Phis are represented as addr * addr list,
even though only the Name constructor of addr will be used.